Refactoring. Changed all the classes to use long instead of int types when referring to player id. Cleaned up code to be less redundant.

+++ src/main/java/mojo/Setup.java
@@ -747,7 +747,7 @@ public class Setup {
         return deck;
     }
 
-    public Player getPlayerById(int id) {
+    public Player getPlayerById(long id) {
         int i = 0;
         for (i = 0; i < playersList.size(); i++) {
