+++ src/main/java/mojo/notification/Notification.java
@@ -6,9 +6,9 @@ package mojo.notification;
  */
 public interface Notification {
 	void setContent();
-	void setReceiver(int playerId);
-	void setSender(int playerId);
-	int getReceiver();
-	int getSender();
+	void setReceiver(long playerId);
+	void setSender(long playerId);
+	long getReceiver();
+	long getSender();
 	void getMessage();
 }
